chore: rename testing module
